Review fixes part 3

+++ PolyEngine/Engine/Src/InputQueue.hpp
@@ -28,15 +28,15 @@ namespace Poly
 		InputEvent(eInputEventType type, eKey key) : Type(type), Key(key) {}
 		InputEvent(eInputEventType type, eMouseButton button) : Type(type), MouseButton(button) {}
 		InputEvent(eInputEventType type, const Vector2i& pos) : Type(type), Pos(pos) {}
-		InputEvent(eInputEventType type, i32 id, eControllerButton button) : Type(type), JoystickID(id), ControllerButton(button) {}
-		InputEvent(eInputEventType type, i32 id, eControllerAxis axis, float value) : Type(type), JoystickID(id), ControllerAxis(axis), AxisValue(value) {}
-		InputEvent(eInputEventType type, i32 id) : Type(type), JoystickID(id) {}
+		InputEvent(eInputEventType type, size_t id, eControllerButton button) : Type(type), JoystickID(id), ControllerButton(button) {}
+		InputEvent(eInputEventType type, size_t id, eControllerAxis axis, float value) : Type(type), JoystickID(id), ControllerAxis(axis), AxisValue(value) {}
+		InputEvent(eInputEventType type, size_t id) : Type(type), JoystickID(id) {}
 
 		eInputEventType Type = eInputEventType::_COUNT;
 		eKey Key = eKey::_COUNT;
 		eMouseButton MouseButton = eMouseButton::_COUNT;
 		Vector2i Pos;
-        i32 JoystickID;
+        size_t JoystickID;
         eControllerButton ControllerButton = eControllerButton::_COUNT;
         eControllerAxis ControllerAxis = eControllerAxis::_COUNT;
         float AxisValue;
